fixup! fix: evaluate expressions from root to support branches
