+++ src/data_factory_testing_framework/functions/evaluator/expression_terminal_transformer.py
@@ -4,7 +4,7 @@
 from lark import Discard, Token, Transformer
 
 from data_factory_testing_framework.exceptions.expression_evaluation_error import ExpressionEvaluationError
-from data_factory_testing_framework.functions.evaluator.rules.expression_rule import EvaluationResult
+from data_factory_testing_framework.functions.evaluator.rules.expression_rule_evaluator import EvaluationResult
 from data_factory_testing_framework.state.run_parameter_type import RunParameterType
