+++ src/data_factory_testing_framework/functions/evaluator/rules/function_call_expression_rule_evaluator.py
@@ -5,7 +5,7 @@ from lark import Tree
 
 from data_factory_testing_framework.exceptions.expression_evaluation_error import ExpressionEvaluationError
 from data_factory_testing_framework.functions.evaluator.exceptions import ExpressionEvaluationInvalidChildTypeError
-from data_factory_testing_framework.functions.evaluator.rules.expression_rule import (
+from data_factory_testing_framework.functions.evaluator.rules.expression_rule_evaluator import (
     EvaluationResult,
     ExpressionRuleEvaluator,
 )
