+++ src/data_factory_testing_framework/functions/evaluator/rules/item_reference_expression_rule_evaluator.py
@@ -4,10 +4,10 @@ from data_factory_testing_framework.exceptions.state_iteration_item_not_set_erro
 from data_factory_testing_framework.functions.evaluator.exceptions import (
     ExpressionEvaluationInvalidNumberOfChildrenError,
 )
-from data_factory_testing_framework.functions.evaluator.rules.expression_rule import EvaluationResult
+from data_factory_testing_framework.functions.evaluator.rules.expression_rule_evaluator import EvaluationResult
 from data_factory_testing_framework.state.pipeline_run_state import PipelineRunState
 
-from .expression_rule import ExpressionRuleEvaluator
+from .expression_rule_evaluator import ExpressionRuleEvaluator
 
 
 class ItemReferenceExpressionRuleEvaluator(ExpressionRuleEvaluator):
