+++ src/data_factory_testing_framework/functions/evaluator/rules/literal_interpolation_expression_rule_evaluator.py
@@ -2,9 +2,9 @@ from lark import Tree
 
 from data_factory_testing_framework.exceptions.expression_evaluation_error import ExpressionEvaluationError
 from data_factory_testing_framework.functions.evaluator.exceptions import ExpressionEvaluationInvalidChildTypeError
-from data_factory_testing_framework.functions.evaluator.rules.expression_rule import EvaluationResult
+from data_factory_testing_framework.functions.evaluator.rules.expression_rule_evaluator import EvaluationResult
 
-from .expression_rule import ExpressionRuleEvaluator
+from .expression_rule_evaluator import ExpressionRuleEvaluator
 
 
 class LiteralInterpolationExpressionRuleEvaluator(ExpressionRuleEvaluator):
