+++ src/data_factory_testing_framework/functions/evaluator/rules/system_variable_reference_expression_rule_evaluator.py
@@ -4,15 +4,16 @@ from data_factory_testing_framework.functions.evaluator.exceptions import (
     ExpressionEvaluationInvalidChildTypeError,
     ExpressionEvaluationInvalidNumberOfChildrenError,
 )
-from data_factory_testing_framework.functions.evaluator.rules.expression_rule import EvaluationResult
+from data_factory_testing_framework.functions.evaluator.rules.expression_rule_evaluator import EvaluationResult
 from data_factory_testing_framework.state.pipeline_run_state import PipelineRunState
+from data_factory_testing_framework.state.run_parameter_type import RunParameterType
 
-from .expression_rule import ExpressionRuleEvaluator
+from .expression_rule_evaluator import ExpressionRuleEvaluator
 
 
-class ActivityReferenceExpressionRuleEvaluator(ExpressionRuleEvaluator):
+class SystemVariableReferenceExpressionRuleEvaluator(ExpressionRuleEvaluator):
     def __init__(self, tree: Tree, state: PipelineRunState) -> None:
-        """Initialize expression rule evaluator."""
+        """Initialize the expression rule."""
         super().__init__(tree)
         self.state: PipelineRunState = state
 
@@ -24,8 +25,11 @@ class ActivityReferenceExpressionRuleEvaluator(ExpressionRuleEvaluator):
                 child_index=0, expected_types=EvaluationResult, actual_type=type(self.children[0])
             )
 
-        self.activity_name = self.children[0].value
+        self.system_variable_name = self.children[0].value
 
     def evaluate(self) -> EvaluationResult:
-        activity_result = self.state.get_activity_result_by_name(self.activity_name)
-        return EvaluationResult(activity_result)
+        system_variable = self.state.get_parameter_by_type_and_name(
+            RunParameterType.System,
+            self.system_variable_name,
+        )
+        return EvaluationResult(system_variable)
