+++ src/data_factory_testing_framework/functions/evaluator/rules/variable_reference_expression_rule_evaluator.py
@@ -4,16 +4,15 @@ from data_factory_testing_framework.functions.evaluator.exceptions import (
     ExpressionEvaluationInvalidChildTypeError,
     ExpressionEvaluationInvalidNumberOfChildrenError,
 )
-from data_factory_testing_framework.functions.evaluator.rules.expression_rule import EvaluationResult
+from data_factory_testing_framework.functions.evaluator.rules.expression_rule_evaluator import EvaluationResult
 from data_factory_testing_framework.state.pipeline_run_state import PipelineRunState
-from data_factory_testing_framework.state.run_parameter_type import RunParameterType
 
-from .expression_rule import ExpressionRuleEvaluator
+from .expression_rule_evaluator import ExpressionRuleEvaluator
 
 
-class SystemVariableReferenceExpressionRuleEvaluator(ExpressionRuleEvaluator):
+class VariableReferenceExpressionRuleEvaluator(ExpressionRuleEvaluator):
     def __init__(self, tree: Tree, state: PipelineRunState) -> None:
-        """Initialize the expression rule."""
+        """Initialize expression rule evaluator."""
         super().__init__(tree)
         self.state: PipelineRunState = state
 
@@ -24,12 +23,8 @@ class SystemVariableReferenceExpressionRuleEvaluator(ExpressionRuleEvaluator):
             raise ExpressionEvaluationInvalidChildTypeError(
                 child_index=0, expected_types=EvaluationResult, actual_type=type(self.children[0])
             )
-
-        self.system_variable_name = self.children[0].value
+        self.variable_name = self.children[0].value
 
     def evaluate(self) -> EvaluationResult:
-        system_variable = self.state.get_parameter_by_type_and_name(
-            RunParameterType.System,
-            self.system_variable_name,
-        )
-        return EvaluationResult(system_variable)
+        variable = self.state.get_variable_by_name(self.variable_name)
+        return EvaluationResult(variable.value)
